requirements.txt
# SPDX-FileCopyrightText: 2022 Renaissance Computing Institute. All rights reserved.
# SPDX-FileCopyrightText: 2023 Renaissance Computing Institute. All rights reserved.
#
# SPDX-License-Identifier: GPL-3.0-or-later
# SPDX-License-Identifier: LicenseRef-RENCI
# SPDX-License-Identifier: MIT

pydantic==1.10.5
fastapi==0.92.0
uvicorn==0.20.0
pyyaml==6.0
psycopg2-binary==2.9.5
pylint==2.16.2